+++ CropsAnytimeAnywhere/release-notes.md
@@ -2,6 +2,8 @@
 
 # Release notes
 ## Upcoming release
+* Added support for tilling any dirt/grass by default.
+* Added options to enable tilling other tile types.
 * Fixed fruit trees not producing fruit outside their normal seasons.
 * Fixed fruit tree textures not updated correctly.